+++ iceoryx_posh/include/iceoryx_posh/internal/mepoo/segment_manager.hpp
@@ -1,5 +1,5 @@
 // Copyright (c) 2019 by Robert Bosch GmbH. All rights reserved.
-// Copyright (c) 2021 by Apex.AI Inc. All rights reserved.
+// Copyright (c) 2021 - 2022 by Apex.AI Inc. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -55,7 +55,7 @@ class SegmentManager
     {
       public:
         SegmentMapping(const ShmName_t& sharedMemoryName,
-                       void* startAddress,
+                       const void* const startAddress,
                        uint64_t size,
                        bool isWritable,
                        uint64_t segmentId,
@@ -71,7 +71,7 @@ class SegmentManager
         }
 
         ShmName_t m_sharedMemoryName{""};
-        void* m_startAddress{nullptr};
+        const void* m_startAddress{nullptr};
         uint64_t m_size{0};
         bool m_isWritable{false};
         uint64_t m_segmentId{0};